Time: 0 ms (100.00%) | Memory: 17 MB (87.46%) - LeetSync

+++ 150-evaluate-reverse-polish-notation/evaluate-reverse-polish-notation.cpp
@@ -11,7 +11,7 @@ public:
             return a/b;
         }
         if(ch=="*"){
-            return (long)a*(long)b;
+            return a*b;
         }
         else return 0;
     }
